Make Thred compile on 64-bit Wine.
Thred can now be compiled to the native Wine architecture of the machine
that compiles it, so there is no more need to install 32-bit
dependencies on 64-bit machines now.
Still needs a few optimisations, as it won't compile without -fpermissive,
but runs fine with the flag set regardless.
Also, the new libhtmlhelp.def file may or may not be usable on 32-bit
machines, still needs testing.

Signed-off-by: Dainius 'GreatEmerald' Masiliunas <pastas4@gmail.com>